#1615 Refactor zoom code into a utility file (for consistency)
